Update article with accordion
