+++ src/styles/App.css
@@ -36,3 +36,7 @@
     transform: rotate(360deg);
   }
 }
+
+.large-text p {
+  font-size: 20px;
+}
